Add initial view for handling race data model.
The race data model currently requires data in the _data/races.yml file.
The structure of the data model requires POSITION and DESCRIPTION
strings for each race as well as a CANDIDATES array.

POSITION is for describing the political office the candidates are
running for.

DESCRIPTION describe the office, the race, and the significance of its
outcome to the active transporation community.

The CANDIDATES array contains a list of candidates that are either being
endorsed outright by BWV or greenlighted. Each candidate has a string
representing their NAME. If the candidate is endorsed, their ENDORSED
boolean is TRUTHY. Otherwise, ENDORSED is FALSEY (and can be ommitted
altogether).

For example,

~~~yaml
- position: President (United States)
  description:
    The President of the United States appoints the Secretary of
    Transportation and influences transporation funding packages. The
    outcome of this race can significantly impact active transporation
    policies at a national scale.
  candidates:
    - name: Frank Underwood
      endorsed: true
    - name: Heather Dunbar
    - name: Henry Mitchell
~~~

This provides data regarding a US Presidential race, giving the position
being run for, a description of the race and its significance, and names
of the candidates greenlighted/endorsed.